[core] Update charts folder structure
