+++ packages/x-charts/src/BarChart/seriesConfig/extremums.ts
@@ -1,4 +1,4 @@
-import { CartesianExtremumGetter } from '../internals/plugins/models/seriesConfig/extremumGetter.types';
+import { CartesianExtremumGetter } from '../../internals/plugins/models/seriesConfig/extremumGetter.types';
 
 const createResult = (data: any, direction: 'x' | 'y') => {
   if (direction === 'x') {
