+++ packages/x-charts/src/LineChart/seriesConfig/index.ts
@@ -1,13 +1,13 @@
-import { ChartSeriesTypeConfig } from '../internals/plugins/models/seriesConfig';
+import { ChartSeriesTypeConfig } from '../../internals/plugins/models/seriesConfig';
 import { getExtremumX, getExtremumY } from './extremums';
-import formatter from './formatter';
+import seriesProcessor from './seriesProcessor';
 import getColor from './getColor';
 import legendGetter from './legend';
 import tooltipGetter from './tooltip';
 
-export const plugin: ChartSeriesTypeConfig<'line'> = {
+export const seriesConfig: ChartSeriesTypeConfig<'line'> = {
   colorProcessor: getColor,
-  seriesProcessor: formatter,
+  seriesProcessor,
   legendGetter,
   tooltipGetter,
   xExtremumGetter: getExtremumX,
