+++ packages/x-charts/src/PieChart/seriesConfig/index.ts
@@ -1,0 +1,12 @@
+import { ChartSeriesTypeConfig } from '../../internals/plugins/models/seriesConfig';
+import seriesProcessor from './seriesProcessor';
+import getColor from './getColor';
+import legendGetter from './legend';
+import tooltipGetter from './tooltip';
+
+export const seriesConfig: ChartSeriesTypeConfig<'pie'> = {
+  colorProcessor: getColor,
+  seriesProcessor,
+  legendGetter,
+  tooltipGetter,
+};
