+++ packages/x-charts/src/PieChart/seriesConfig/seriesProcessor.ts
@@ -1,9 +1,9 @@
 import { pie as d3Pie } from '@mui/x-charts-vendor/d3-shape';
-import { ChartSeriesDefaultized } from '../models/seriesType/config';
-import { ChartsPieSorting, PieValueType } from '../models/seriesType/pie';
-import { SeriesId } from '../models/seriesType/common';
-import { getLabel } from '../internals/getLabel';
-import { SeriesProcessor } from '../internals/plugins/models';
+import { ChartSeriesDefaultized } from '../../models/seriesType/config';
+import { ChartsPieSorting, PieValueType } from '../../models/seriesType/pie';
+import { SeriesId } from '../../models/seriesType/common';
+import { getLabel } from '../../internals/getLabel';
+import { SeriesProcessor } from '../../internals/plugins/models';
 
 const getSortingComparator = (comparator: ChartsPieSorting = 'none') => {
   if (typeof comparator === 'function') {
@@ -21,7 +21,7 @@ const getSortingComparator = (comparator: ChartsPieSorting = 'none') => {
   }
 };
 
-const formatter: SeriesProcessor<'pie'> = (params) => {
+const seriesProcessor: SeriesProcessor<'pie'> = (params) => {
   const { seriesOrder, series } = params;
 
   const defaultizedSeries: Record<SeriesId, ChartSeriesDefaultized<'pie'>> = {};
@@ -61,4 +61,4 @@ const formatter: SeriesProcessor<'pie'> = (params) => {
   };
 };
 
-export default formatter;
+export default seriesProcessor;
