+++ packages/x-charts/src/ScatterChart/ScatterPlot.tsx
@@ -2,7 +2,7 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import { Scatter, ScatterProps } from './Scatter';
-import getColor from './getColor';
+import getColor from './seriesConfig/getColor';
 import { useScatterSeries } from '../hooks/useSeries';
 import { useXAxes, useYAxes } from '../hooks';
 import { useZAxes } from '../hooks/useZAxis';
