+++ packages/x-charts/src/ScatterChart/seriesConfig/index.ts
@@ -1,12 +1,12 @@
-import { ChartSeriesTypeConfig } from '../internals/plugins/models/seriesConfig';
+import { ChartSeriesTypeConfig } from '../../internals/plugins/models/seriesConfig';
 import { getExtremumX, getExtremumY } from './extremums';
-import formatter from './formatter';
+import seriesProcessor from './seriesProcessor';
 import getColor from './getColor';
 import legendGetter from './legend';
 import tooltipGetter from './tooltip';
 
-export const plugin: ChartSeriesTypeConfig<'scatter'> = {
-  seriesProcessor: formatter,
+export const seriesConfig: ChartSeriesTypeConfig<'scatter'> = {
+  seriesProcessor,
   colorProcessor: getColor,
   legendGetter,
   tooltipGetter,
