+++ packages/x-charts/src/context/ChartProvider/ChartProvider.tsx
@@ -12,17 +12,17 @@ import { useChartCartesianAxis } from '../../internals/plugins/featurePlugins/us
 import { useChartInteraction } from '../../internals/plugins/featurePlugins/useChartInteraction';
 import { useChartZAxis } from '../../internals/plugins/featurePlugins/useChartZAxis';
 import { useChartHighlight } from '../../internals/plugins/featurePlugins/useChartHighlight/useChartHighlight';
-import { plugin as barPlugin } from '../../BarChart/plugin';
-import { plugin as scatterPlugin } from '../../ScatterChart/plugin';
-import { plugin as linePlugin } from '../../LineChart/plugin';
-import { plugin as piePlugin } from '../../PieChart/plugin';
+import { seriesConfig as barSeriesConfig } from '../../BarChart/seriesConfig';
+import { seriesConfig as scatterSeriesConfig } from '../../ScatterChart/seriesConfig';
+import { seriesConfig as lineSeriesConfig } from '../../LineChart/seriesConfig';
+import { seriesConfig as pieSeriesConfig } from '../../PieChart/seriesConfig';
 import { ChartSeriesType } from '../../models/seriesType/config';
 
 export const defaultSeriesConfig: ChartSeriesConfig<'bar' | 'scatter' | 'line' | 'pie'> = {
-  bar: barPlugin,
-  scatter: scatterPlugin,
-  line: linePlugin,
-  pie: piePlugin,
+  bar: barSeriesConfig,
+  scatter: scatterSeriesConfig,
+  line: lineSeriesConfig,
+  pie: pieSeriesConfig,
 };
 
 // For consistency with the v7, the cartesian axes are set by default.
